Use safe default for local speech transcription

+++ src/openjarvis/core/config.py
@@ -1217,7 +1217,7 @@ class SpeechConfig:
     model: str = "base"  # Whisper model size: tiny, base, small, medium, large-v3
     language: str = ""  # Empty = auto-detect
     device: str = "auto"  # "auto", "cpu", "cuda"
-    compute_type: str = "float16"  # "float16", "int8", "float32"
+    compute_type: str = "int8"  # "int8", "float16", "float32"
 
 
 @dataclass(slots=True)
