+++ tests/speech/test_config.py
@@ -9,7 +9,7 @@ def test_speech_config_defaults():
     assert cfg.model == "base"
     assert cfg.language == ""
     assert cfg.device == "auto"
-    assert cfg.compute_type == "float16"
+    assert cfg.compute_type == "int8"
 
 
 def test_jarvis_config_has_speech():
